requirements.txt
# Telegram Bot (версия 21+ для совместимости с Python 3.13)
python-telegram-bot>=21.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests для Hugging Face API
requests>=2.31.0

# Database
sqlalchemy>=2.0.0
pymongo>=4.6.0

# Date/Time utilities
python-dateutil>=2.8.0

# Image Processing (совместимо с Python 3.13)
Pillow>=10.2.0

# Optional: для будущих фич
# matplotlib>=3.8.0  # для визуализации timeline
# reportlab>=4.0.0   # для генерации PDF
# schedule>=1.2.0    # для напоминаний

